+++ utils/productApi.js
@@ -1,7 +1,7 @@
 // productApi.js
-export async function fetchProducts({ pageSize, page, keyword, sortBy }) {
+export async function fetchProducts({ pageSize, page, keyword, orderBy }) {
   const response = await fetch(
-    `https://panda-market-api.vercel.app/products?pageSize=${pageSize}&page=${page}&keyword=${keyword}&sortBy=${sortBy}`
+    `https://panda-market-api.vercel.app/products?pageSize=${pageSize}&page=${page}&keyword=${keyword}&orderBy=${orderBy}`
   );
   if (!response.ok) {
     throw new Error("Failed to fetch products");
